+++ lib/fledex/animation/coordinator.ex
@@ -72,10 +72,10 @@ defmodule Fledex.Animation.Coordinator do
   """
   @spec start_link(strip_name :: atom, coordinator_name :: atom, config :: config_t) ::
           GenServer.on_start()
-  def start_link(strip_name, animation_name, config) do
+  def start_link(strip_name, coordinator_name, config) do
     {:ok, _pid} =
-      GenServer.start_link(__MODULE__, {strip_name, animation_name, config},
-        name: Utils.via_tuple(strip_name, :coordinator, animation_name)
+      GenServer.start_link(__MODULE__, {strip_name, coordinator_name, config},
+        name: Utils.via_tuple(strip_name, :coordinator, coordinator_name)
       )
   end