+++ lib/fledex/animation/manager.ex
@@ -40,8 +40,6 @@ defmodule Fledex.Animation.Manager do
         }
 
   @typep state_t :: %{
-           #           animations: %{atom => Animator.config_t()},
-           coordinators: %{atom => Coordinator.config_t()},
            jobs: %{atom => JobScheduler.config_t()}
          }
 
@@ -125,8 +123,6 @@ defmodule Fledex.Animation.Manager do
     Process.flag(:trap_exit, true)
 
     state = %{
-      # animations: %{},
-      coordinators: %{},
       jobs: %{}
     }
 
@@ -142,7 +138,7 @@ defmodule Fledex.Animation.Manager do
           {:reply, :ok, state_t}
   def handle_call({:register_strip, strip_name, drivers, strip_config}, _pid, state)
       when is_atom(strip_name) do
-    if led_strip_registered?(strip_name, state) do
+    if AnimationSystem.led_strip_exists?(strip_name) do
       # we have a bit of a problem when using the kino driver, since it will not be reinitialized
       # when calling this function again (and thereby we don't get any frame/display).
       # Therefore we add here an extra step to reinitiate the the drivers while registering the strip.
@@ -179,7 +175,7 @@ defmodule Fledex.Animation.Manager do
   @spec terminate(atom, state_t) :: :ok
   def terminate(_reason, state) do
     Logger.debug("shutting down Animation.Manager")
-    strip_names = Map.keys(state.coordinators)
+    strip_names = AnimationSystem.get_led_strips()
 
     _state =
       Enum.reduce(strip_names, state, fn strip_name, state ->
@@ -188,8 +184,6 @@ defmodule Fledex.Animation.Manager do
 
     :ok
   end
-
-  ### MARK: public helper functions
 
   ### MARK: private helper fucntions
   # we split the "animation" into the different aspects
@@ -208,9 +202,6 @@ defmodule Fledex.Animation.Manager do
       raise RuntimeError, "An unknown type was encountered #{inspect(rest)}"
     end
 
-    # # animations = Enum.filter(config, fn {_key, value} -> value.type in [:animation, :static] end)
-    # # coordinators = Enum.filter(config, fn {_key, value} -> value.type in [:coordinator] end)
-    # # jobs = Enum.filter(config, fn {_key, value} -> value.type in [:job] end)
     {animations, coordinators, jobs}
   end
 
@@ -218,33 +209,19 @@ defmodule Fledex.Animation.Manager do
   defp register_strip(state, strip_name, drivers, strip_config) do
     _result = AnimationSystem.start_led_strip(strip_name, drivers, strip_config)
 
-    %{
-      state
-      | # animations: Map.put_new(state.animations, strip_name, nil),
-        coordinators: Map.put_new(state.coordinators, strip_name, nil),
-        jobs: Map.put_new(state.jobs, strip_name, nil)
-    }
+    %{state | jobs: Map.put_new(state.jobs, strip_name, nil)}
   end
 
   @spec unregister_strip(state_t, atom) :: state_t
   defp unregister_strip(state, strip_name) do
     # Logger.info("unregistering led_strip_ #{strip_name}")
-    shutdown_coordinators(
-      strip_name,
-      Map.keys(state.coordinators[strip_name] || %{})
-    )
-
+    shutdown_coordinators(strip_name, LedStripSupervisor.get_coordinators(strip_name))
     shutdown_jobs(strip_name, Map.keys(state.jobs[strip_name] || %{}))
     shutdown_animators(strip_name, LedStripSupervisor.get_animations(strip_name))
-    AnimationSystem.stop_led_strip(strip_name)
-    # :ok = LedStripSupervisor.stop(strip_name)
 
-    %{
-      state
-      | # animations: Map.drop(state.animations, [strip_name]),
-        coordinators: Map.drop(state.coordinators, [strip_name]),
-        jobs: Map.drop(state.coordinators, [strip_name])
-    }
+    AnimationSystem.stop_led_strip(strip_name)
+
+    %{state | jobs: Map.drop(state.jobs, [strip_name])}
   end
 
   @spec register_animations(state_t, atom, map) :: state_t
@@ -263,7 +240,6 @@ defmodule Fledex.Animation.Manager do
     update_animators(strip_name, updated)
     create_animators(strip_name, created)
 
-    # %{state | animations: Map.put(state.animations, strip_name, configs)}
     state
   end
 
@@ -271,7 +247,6 @@ defmodule Fledex.Animation.Manager do
   defp shutdown_animators(strip_name, dropped_animations) do
     Enum.each(dropped_animations, fn animation_name ->
       LedStripSupervisor.stop_animation(strip_name, animation_name)
-      # :ok = Animator.stop(strip_name, animation_name)
     end)
   end
 
@@ -342,13 +317,13 @@ defmodule Fledex.Animation.Manager do
   @spec register_coordinators(state_t, atom, map) :: state_t
   defp register_coordinators(state, strip_name, coordinators) do
     {dropped, updated, created} =
-      filter_configs(Map.get(state.coordinators, strip_name), coordinators)
+      filter_configs2(LedStripSupervisor.get_coordinators(strip_name), coordinators)
 
     shutdown_coordinators(strip_name, dropped)
     update_coordinators(strip_name, updated)
     create_coordinators(strip_name, created)
 
-    %{state | coordinators: Map.put(state.coordinators, strip_name, coordinators)}
+    state
   end
 
   @spec create_coordinators(atom, map) :: :ok
@@ -368,7 +343,7 @@ defmodule Fledex.Animation.Manager do
   @spec shutdown_coordinators(atom, [atom]) :: :ok
   defp shutdown_coordinators(strip_name, coordinator_names) do
     Enum.each(coordinator_names, fn coordinator_name ->
-      :ok = Coordinator.stop(strip_name, coordinator_name)
+      LedStripSupervisor.stop_coordinator(strip_name, coordinator_name)
     end)
   end
 
@@ -408,10 +383,5 @@ defmodule Fledex.Animation.Manager do
 
       if Keyword.get(job_config.options, :run_once, false), do: JobScheduler.run_job(job)
     end)
-  end
-
-  @spec led_strip_registered?(atom, state_t) :: boolean
-  defp led_strip_registered?(strip_name, _state) when is_atom(strip_name) do
-    AnimationSystem.led_strip_exists?(strip_name)
   end
 end