+++ lib/fledex/supervisor/utils.ex
@@ -6,6 +6,10 @@ defmodule Fledex.Supervisor.Utils do
   @moduledoc """
   Some utilities related to the animatino system.
   """
+
+  alias Fledex.Animation.Animator
+  alias Fledex.Animation.Coordinator
+
   @pubsub Mix.Project.config()[:app]
   @app_supervisor Fledex.DynamicSupervisor
   @registry Fledex.Supervisor.WorkersRegistry
@@ -42,4 +46,61 @@ defmodule Fledex.Supervisor.Utils do
   @spec via_tuple(atom, :animator | :job | :coordinator | :led_strip, atom) :: GenServer.name()
   def via_tuple(strip_name, type, animation_name),
     do: {:via, Registry, {@registry, {strip_name, type, animation_name}}}
+
+  @spec supervisor_name(atom) :: GenServer.name()
+  def supervisor_name(strip_name) do
+    via_tuple(strip_name, :led_strip, :supervisor)
+  end
+
+  @spec workers_name(atom) :: GenServer.name()
+  def workers_name(strip_name) do
+    via_tuple(strip_name, :led_strip, :workers)
+  end
+
+  @spec start_worker(
+          atom,
+          atom,
+          Animator | Coordinator,
+          Animator.config_t() | Coordinator.config_t()
+        ) :: DynamicSupervisor.on_start_child()
+  def start_worker(strip_name, name, type, config) do
+    DynamicSupervisor.start_child(
+      workers_name(strip_name),
+      %{
+        # no need to be unique
+        id: name,
+        start: {type, :start_link, [strip_name, name, config]},
+        restart: :transient
+      }
+    )
+  end
+
+  @spec worker_exists?(atom, :coordinator | :animator | :led_strip, atom) :: boolean
+  def worker_exists?(strip_name, type, name) do
+    case Registry.lookup(worker_registry(), {strip_name, type, name}) do
+      [{_pid, _value}] -> true
+      _other -> false
+    end
+  end
+
+  @spec get_workers(atom, :coordinator | :animator | :led_strip, atom) :: list(atom)
+  def get_workers(strip_name, type, name) do
+    Registry.select(worker_registry(), [
+      {
+        {{strip_name, type, name}, :_, :_},
+        [],
+        [:"$1"]
+      }
+    ])
+  end
+
+  @spec stop_worker(GenServer.name(), atom, :coordinator | :animator | :led_strip, atom) :: :ok
+  def stop_worker(supervisor, strip_name, type, name) do
+    case Registry.lookup(worker_registry(), {strip_name, type, name}) do
+      [{pid, _value}] -> DynamicSupervisor.terminate_child(supervisor, pid)
+      _other -> :ok
+    end
+
+    :ok
+  end
 end